Improve performance when handling comments on stop_times
Previoulsy the copy to invert the stop_time_ids was done for every
comments, so it's a bit slow. This do it only one time to speed it.

This can be improved by doing the copy lazily to not pay the cost if
there isn't any comments on stop_times